+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Change Log
   
 All notable changes to the 'vscode-anchor' extension will be documented in this file.  
+
+## Version 0.2.4 [current]  
+### Added  
+- Support for Anchor Version Manager (avm)
+- Scaffolding projects Improvements #12
+- Test against local validator
   
-## Version 0.2.3 [current]  
+### Changed  
+- Install command now uses avm to install Anchor and Other versions.  
+- Edit buttons on views for programs and tests are now replaced with default name click to open file.
+  
+### Fixed  
+- Anchor tests were not visible in pane #13
+  
+---
+## Version 0.2.3 
 ### Added  
 - Added Help Commands to your favourite docs site.
 - Analyze your code with Soteria, Cargo Audit, and clippy.